agent_core/requirements.txt
grpcio==1.59.0
grpcio-tools==1.59.0
protobuf==4.24.4
python-dotenv==1.0.0
psutil==5.9.5
abacusai==1.0.0
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0
pytest-xdist==3.5.0
